requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.6.3
Deprecated==1.2.14
distlib==0.3.8
distro==1.8.0
dj-database-url==2.2.0
Django==3.2.25
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-heroku==0.3.1
django-social-share==2.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dnspython==2.5.0
elevenlabslib==0.18.0
email-validator==2.1.0.post1
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-redis==0.4.0
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.12.2
future==1.0.0
fuzzywuzzy==0.18.0
getmac==0.9.4
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
instagrapi==2.0.3
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
Levenshtein==0.23.0
llama-index==0.9.34
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.25.2
openai==1.9.0
opencv-python==4.8.1.78
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
platformdirs==4.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyMySQL==1.1.0
pyodbc==5.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.23.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.11.4
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
stripe==10.1.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.25.1
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.7.0
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.3
